+++ lib/binance_client/responses/sub_account_set_bnb_burn_status_response.rb
@@ -1,0 +1,13 @@
+module BinanceClient
+  class SubAccountSetBnbBurnStatusResponse < BaseResponse
+
+    def sub_account_id
+      body["subAccountId"]
+    end
+
+    def spot_bnb_burn?
+      body["spotBNBBurn"]
+    end
+
+  end
+end